pyproject.toml
[project]
name = "uba-ai-auto-generate-questionary-api"
version = "0.0.1"
description = ""
authors = [
    {name = "Pedro Vieira",email = "pedro515vieira@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "api-crud-generate-libary @ git+https://github.com/pedro-vlps/api_crud_generate_libary.git",
    "openai (>=2.32.0,<3.0.0)",
    "uvicorn (>=0.44.0,<0.45.0)",
    "sqlacodegen (>=4.0.3,<5.0.0)",
    "pydantic-settings (>=2.14.0,<3.0.0)",
    "cryptography (>=42.0.0,<43.0.0)",
    "psycopg2-binary (>=2.9.12,<3.0.0)",
    "pyjwt (>=2.12.1,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
